Library Basics (#27)
This PR sets up library basics. It includes implementation, documentation, tests, and benchmarks for the following modules:

- Archives
  - GridArchive
  - CVTArchive
  - SlidingBoundariesArchive
- Emitters
  - GaussianEmitter
  - IsoLineEmitter
- Optimizers
  - Optimizer
- Factory
- Visualize

We have also set up the documentation pages, though they are WIP and will be addressed in followup PRs. We have also set up many of the examples, though most of them are still WIP as well.

Co-authored-by: Matthew Fontaine
Co-authored-by: Samuel Sommerer <sommerer@usc.edu>
Co-authored-by: Yulun Zhang <yulunzha@usc.edu>