Rename elites_with_measures to retrieve (#295)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

`elites_with_measures` and `elites_with_measures_single` is a very long
name. `retrieve` is much more succinct.

While the name may be more vague, we think this is okay because we
review the method in the lunar lander tutorial. Furthermore, the usage
makes the meaning very clear; e.g., `elite_batch =
archive.retrieve(measures_batch)`.

We also considered the name `find` but `find` implies a search, which is
not what this method is doing. This method is merely _retrieving_ the
elite at a given cell, not searching the archive to find an elite with
the given measures.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in
[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in `HISTORY.md`
- [x] This PR is ready to go