Schedulers warn if no solutions are inserted into archive (#320)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Solutions may not be inserted into the archive if the threshold_min is
too high. This can lead to unexpected behavior. For instance,
GaussianEmitter with an initial set of solutions will keep producing
that same set of solutions, and archive metrics like `obj_max` will be
`None`. We shouldn't throw an error when this happens as users may
decide they want this, but a warning is nice.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Warn in Scheduler
- [x] Warn in BanditScheduler
- [x] Test Scheduler warning
- [x] Test BanditScheduler warning

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in
[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in `HISTORY.md`
- [x] This PR is ready to go