Add visualization of QDax repertoires (#353)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

QDax’s repertoires are not too different from pyribs in terms of the
data stored, which means our visualization tools can easily be used for
both libraries. This PR makes it possible to visualize one of the main
repertoires used in QDax, i.e., the MAPElitesRepertoire. The key idea is
to transform the MAPElitesRepertoire into a CVTArchive and then plot the
result with `cvt_archive_heatmap`. In this PR, this functionality is
exposed via a new function, `ribs.visualize.qdax_repertoire_heatmap`. We
update the tests and CI accordingly.

Note 1: This PR does not introduce a dependency on QDax for pyribs. If a
QDax repertoire is passed into the method introduced here, it can be
plotted, but we do not import qdax itself at any point.

Note 2: QDax does not have a separate grid archive. Instead, both grid
and CVT archives from pyribs are represented with the
MAPElitesRepertoire. Specifically, one can choose Euclidean centroids
when constructing the repertoire in order to create a "grid archive."

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x]  Write/document function
- [x]  Add tests in tests/visualize_qdax
- [x]  Set up CI for testing qdax visualization

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go