Allow using kwargs for colorbar in parallel_axes_plot (#358)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This PR makes the colorbar usage more consistent with the other
visualization methods. We now have a `cbar` arg which determines whether
the colorbar is plotted. We also have a `cbar_kwargs` arg which allows
configuration of the colorbar. The `cbar_kwargs` has sensible defaults
for `orientation` and `pad` that align with the previous default values
for `cbar_orientation` and `cbar_pad`.

Note that this is a breaking change because it removes
`cbar_orientation` and `cbar_pad`.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Add `cbar` and `cbar_kwargs` args to parallel_axes_plot
- [x] Remove `cbar_pad` and `cbar_orientation` args
- [x] Update tests

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go