Refactor argument validation utilities (#365)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Currently, the argument validation utilities do not perform
preprocessing on the arguments passed into them. However, every call of
these utilities is prefaced with preprocessing (e.g., converting with
asarray). This PR moves the preprocessing into the utilities. This
should make it easier to handle argument validation as there is no
longer a need to repeat preprocessing code.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Refactor validate_batch_args
- [x] Refactor validate_single_args

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go