Fix lunar lander dependency issues (#367)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

We have been running into some issues with lunar lander not running
properly on Colab due to dependency issues. We also had some dependency
issues with swig (see #366).

Regarding swig, I found that it was difficult to get the examples to
work properly with swig because we were not using a virtual environment,
so the different Python versions were being confused -- hence we would
get errors like "Module not found" when trying to use swig. To remedy
this problem, I have started #367 to run all PR jobs in a Conda env.
This PR uses a Conda env only for the examples job.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Upgrade to gymnasium 0.29.1 due to Jax-jumpy issues -> gymnasium
0.27.0 depended on jax-jumpy, which resulted in installation errors
because jax-jumpy was not fully production-ready
- [x]  Fix ordering of Dask imports
- [x] Re-run tutorial on Colab to get times; update tutorial times
accordingly
- [x] Run on Colab and fix deprecation warnings -> looks like there were
no warnings
- [x] Try running locally in Python 3.10 -> Colab CPU seems really slow
compared to local CPU, but this makes sense
- [x]  Fix deps in lunar lander example
- [x] Make example.sh script run multiple pip install commands

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go