Use Conda envs in all CI jobs (#368)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

We encountered an issue in #366 where different Python instances caused
swig to not be found. To prevent this and similar issues in the future,
this PR creates Conda envs to ensure that we always / are more likely to
use the correct version of Python.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

## Questions

<!-- Any concerns or points of confusion? -->

- [x] Should we use Tox?
- For now, I believe Tox would introduce too much complexity. All we
want is to create a virtual env for each job. Tox may be able to do this
with less boilerplate, but we would also incur overhead to learn how to
use it properly. Conda envs also more closely match our audience of
researchers.
- Note that we did have Tox originally, but we removed it in #143 to
simplify our workflows.

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go