Import ribs[visualize] in tutorials that need it (#379)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Since the visualize extra is installed in the CI for the tutorials, we
did not detect bugs where the tutorials needed ribs[visualize] but only
installed ribs. This PR adds ribs[visualize] to the tutorials that need
it and also makes a note in the "Adding a Tutorial" section of
CONTRIBUTING.md.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go