Switch to a branch-based release model (#382)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Pyribs roughly follows a trunk-based development model. Previously, for
releases, we only released from the trunk (master branch). However, this
does not work well if we need to perform a patch release while trunk
already has several features implemented for the next release. This PR
shifts to the [branch-based release
model](https://trunkbaseddevelopment.com/branch-for-release/) where we
create release branches for minor versions. If we need to release
patches, we can cherry pick commits from trunk into the release branch
and then push the patch release. We will use this to properly release
0.6.2, as 0.6.1 was released with features for 0.7.0 (namely, OMG-MEGA
and OG-MAP-Elites).

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Update instructions in CONTRIBUTING.md
- [x] Update HISTORY.md

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go