Test pyribs installation in tutorials (#384)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, we always installed `ribs[visualize]` before testing
tutorials. This led to uncaught bugs where some tutorials needed
`ribs[visualize]` to be installed but did not install it -- namely,
since we always installed `ribs[visualize]`, the tutorial could get by
without installing `ribs[visualize]`. This led to bugs such as those in
#379. To remedy this issue, this PR removes the default installation of
`ribs[visualize]` in the tutorial CI.

Instead, the notebook now installs the local copy of pyribs on its own
-- we do this by replacing, e.g., `pip install ribs[visualize]` with
`pip install .[visualize]`. Note that it is not possible to leave this
dependency as is (i.e., `pip install ribs[visualize]`) because the
tutorial would pull from PyPI instead of the local copy.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Fix bug in fooling mnist tutorial caused by matplotlib not being
installed

## Questions

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go