Make ArrayStore fields be valid identifiers (#398)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

Although we mostly use the field names as dict keys, which can take on
any value, it may be useful in the future to be able to use fields as
identifiers. For example, if we are using kwargs, it makes sense to be
able to pass in a field name as `f(field1=[...])`, which is not possible
if `field1` is an invalid identifier like `field foo` (note the space).

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Add check
- [x] Add test

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go