Add field_desc property to ArrayStore (#403)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This property is a cached property. It takes some computation, but we
only want to do it once since it never changes. Furthermore, we do not
want to store it in the constructor since we would then need to worry
about it when reconstructing in `from_raw_dict`.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go