Reimplement ArchiveBase using ArrayStore (#399)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

This PR refactors the ArchiveBase to place data in ArrayStore rather
than in separate NumPy arrays. This should make the archive more
flexible and easier to extend in the future.

While flexibility is the end goal, note that this PR is only a refactor;
thus, no API changes are being made. As a result, tests will mostly
remain the same.

The exception is the creation of a private `ribs.archives._transforms`
module, for which we do introduce new tests. This module should be
considered unstable for now, but once it seems more stable, we can make
it public.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Modify ArchiveBase
- [x] Update SlidingBoundariesArchive
- [x] Create private transforms module
- [x] Test transforms
- [x] Fix ArchiveBase tests where appropriate

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog
in
      `HISTORY.md`
- [x] This PR is ready to go